Move monitor command to its own package
